+++ src/app/services/sleep-detection-automations/sleep-mode-change-on-steamvr-status-automation.service.ts
@@ -1,0 +1,51 @@
+import { Injectable } from '@angular/core';
+import { AutomationConfigService } from '../automation-config.service';
+import { listen } from '@tauri-apps/api/event';
+import { OpenVRService } from '../openvr.service';
+import {
+  AUTOMATION_CONFIGS_DEFAULT,
+  AutomationType,
+  SleepModeChangeOnSteamVRStatusAutomationConfig,
+  SleepModeDisableAtTimeAutomationConfig,
+} from '../../models/automations';
+import { cloneDeep } from 'lodash';
+import { debounceTime, filter, map, pairwise, tap } from 'rxjs';
+import { SleepService } from '../sleep.service';
+import { SleepModeStatusChangeReasonBase } from '../../models/sleep-mode';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class SleepModeChangeOnSteamVRStatusAutomationService {
+  private config: SleepModeChangeOnSteamVRStatusAutomationConfig = cloneDeep(
+    AUTOMATION_CONFIGS_DEFAULT.SLEEP_MODE_CHANGE_ON_STEAMVR_STATUS
+  );
+
+  constructor(
+    private automationConfig: AutomationConfigService,
+    private openvr: OpenVRService,
+    private sleep: SleepService
+  ) {}
+
+  async init() {
+    this.automationConfig.configs
+      .pipe(map((configs) => configs.SLEEP_MODE_CHANGE_ON_STEAMVR_STATUS))
+      .subscribe((config) => (this.config = config));
+
+    this.openvr.status
+      .pipe(
+        map((status) => status === 'INITIALIZED'),
+        debounceTime(2000),
+        pairwise(),
+        tap(([initializedBefore, initializedAfter]) => {
+          if (initializedBefore && !initializedAfter && this.config.disableOnSteamVRStop) {
+            this.sleep.disableSleepMode({
+              type: 'AUTOMATION',
+              automation: 'SLEEP_MODE_CHANGE_ON_STEAMVR_STATUS',
+            });
+          }
+        })
+      )
+      .subscribe();
+  }
+}